Refactor to provider-based plugin (extensibility)
Existing Nyaa functionality is now the only built-in provider, serving
as an example for supporting other sites.

Anyone can now publish a package with (a) `sopel_torrentinfo.providers`
entry point(s) that handles links for any arbitrary site.

Of course there is no actual limitation to torrent indexes. Honestly,
this approach looks great for a generic, pluggable link-info fetcher
usable with any site. The plugin handles fetch errors for free, and the
provider/entry-point deals with parsing the data if successful.