Add URL for non-view Nyaa links & modify provider interface to support
Requires passing the `trigger` to `TorrentInfoProvider.parse()` calls,
since it must inspect the match groups to decide whether a link should
be part of the output or not.

Reformatted the messy `bot.say()` in the main plugin file, since I had
to touch it (and make it even longer) anyway.